requirements.txt
Django==4.2.6
Pillow==10.0.1
django-admin-sortable2==2.1.10